Add diagnose SNR measurement and fix ShieldWorldArchiver FP32 precision
ShieldWorldArchiver: remove .half() downcast from SVD factors so EF
buffers at small magnitudes (1e-7) survive checkpoint roundtrip without
underflow.  Decompression .float() calls kept for legacy FP16 compat.

diagnose(): add _compute_snr_db() helper that simulates full TCFP-12
dual decomposition (W_hi + W_lo) roundtrip and reports per-layer SNR
in dB.  Gated by sample_batch parameter.  Warns when any layer < 30 dB.
Guards zero-signal (nan) and zero-noise (inf) edge cases.  Aggregates
exclude inf/nan for meaningful min/mean statistics.

8 new tests (5 diagnose SNR + 3 ShieldWorldArchiver), 428 total passing.

+++ tcfp/training/checkpointing.py
@@ -360,9 +360,9 @@ class ShieldWorldArchiver:
                 rank = max(1, int(min(t.shape) * rank_fraction))
                 try:
                     U, S, Vt = torch.linalg.svd(t, full_matrices=False)
-                    compressed_buffers[f"{key}.U"] = U[:, :rank].half()
-                    compressed_buffers[f"{key}.S"] = S[:rank].half()
-                    compressed_buffers[f"{key}.Vt"] = Vt[:rank, :].half()
+                    compressed_buffers[f"{key}.U"] = U[:, :rank]
+                    compressed_buffers[f"{key}.S"] = S[:rank]
+                    compressed_buffers[f"{key}.Vt"] = Vt[:rank, :]
                 except Exception:
                     # Fall back to uncompressed on SVD failure
                     compressed_buffers[key] = tensor
